Use poetry for packaging and dependency management
Use [poetry](https://python-poetry.org/) as a packaging and dependency
management tool. It replaces `requirements*.{in,txt}` files with
`pyproject.toml` (the new standard file for python project
configuration) and `poetry.lock` (lock file, specific to poetry.)

Poetry also automatically creates per-project virtual environements, so
it can replace `tox` as well.

This config defines a python package that can be published to pypi,
named `sigsum-witness`, so that we can eventually offer
`pip install sigsum-witness` as an installation method.

To be able to installs a `sigsum-witness` script with poetry,
`sigsum-witness.py` was renamed `sigsum_witness.py`.